refactor(all): Move to @accordproject scope
