+++ packages/concerto-ui-core/babel.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    "presets": [
+       ['@babel/preset-env', { targets: { node: 'current' } }]
+    ],
+    "env": {
+      "test": {
+        "plugins": [
+          "@babel/plugin-transform-modules-commonjs"
+        ]
+      }
+    }
+  };
