+++ packages/concerto-ui-core/package.json
@@ -5,30 +5,22 @@
     "composer-concerto": "^0.70.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "rm -rf dist && NODE_ENV=production babel src/lib --out-dir dist --copy-files --ignore __tests__,spec.js,test.js,__snapshots__",
-    "test": "CI=true react-scripts test",
-    "eject": "react-scripts eject",
-    "postlint": "npm run licchk",
+    "build": "rm -rf dist && NODE_ENV=production babel src --out-dir dist --ignore '**/*.spec.js,**/*.test.js,samples,__snapshots__,**/*.stories.js'",
+    "watch": "npm run build -- --watch",
+    "test": "jest --coverage --verbose --color",
+    "pretest": "npm run licchk",
     "licchk": "license-check-and-add",
     "prepublishOnly": "npm run build"
   },
   "eslintConfig": {
     "extends": "react-app"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
   "devDependencies": {
     "@babel/cli": "7.4.3",
     "@babel/core": "7.4.3",
-    "license-check-and-add": "2.3.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "2.1.8"
+    "babel-jest": "24.7.1",
+    "jest": "24.7.1",
+    "license-check-and-add": "2.3.1"
   },
   "main": "dist/index.js",
   "module": "dist/index.js",
@@ -36,6 +28,11 @@
     "dist",
     "README.md"
   ],
+  "jest": {
+    "setupFiles": [
+      "../../setupTests"
+    ]
+  },
   "license-check-and-add-config": {
     "folder": ".",
     "license": "LICENSE.txt",
