--- packages/concerto-ui-react-demo/.babelrc
@@ -1,3 +1,0 @@
-{
-    "presets": ["env", "react","transform-runtime"]
-}
