+++ packages/concerto-ui-react-demo/.gitignore
@@ -43,8 +43,9 @@ bower_components
 # node-waf configuration
 .lock-wscript
 
-# Compiled binary addons (https://nodejs.org/api/addons.html)
-build/Release
+# Build
+dist
+build
 
 # Dependency directories
 node_modules/
