+++ packages/concerto-ui-react-demo/config-overrides.js
@@ -1,0 +1,21 @@
+let path = require ('path');
+let fs = require ('fs');
+const {
+  override,
+  addDecoratorsLegacy,
+  babelInclude,
+  disableEsLint,
+} = require('customize-cra');
+
+module.exports = function (config, env) {
+  return Object.assign(config, override(
+    disableEsLint(),
+    addDecoratorsLegacy(),
+    /*Make sure Babel compiles the stuff in the common folder*/
+    babelInclude([
+      path.resolve('src'), // don't forget this
+      fs.realpathSync('node_modules/@accordproject/concerto-ui-react/src')
+    ])
+  )(config, env)
+  );
+};
